Process null check

+++ src/WebJobs.Script/Workers/ProcessManagement/WorkerProcess.cs
@@ -214,7 +214,7 @@ namespace Microsoft.Azure.WebJobs.Script.Workers
 
                 if (Process != null)
                 {
-                    if (!Process.HasExited)
+                    if (Process != null && !Process.HasExited)
                     {
                         Process.Kill();
                         if (!Process.WaitForExit(processExitTimeoutInMilliseconds))
@@ -222,7 +222,10 @@ namespace Microsoft.Azure.WebJobs.Script.Workers
                             _workerProcessLogger.LogWarning($"Worker process has not exited despite waiting for {processExitTimeoutInMilliseconds} ms");
                         }
                     }
-                    Process.Dispose();
+                    if (Process != null)
+                    {
+                        Process.Dispose();
+                    }
                 }
             }
             catch (Exception exc)
